+++ init.lua
@@ -67,3 +67,5 @@ vim.o.backup = false
 vim.o.writebackup = false
 vim.o.swapfile = false
 
+-- 相对数字
+vim.o.relativenumber = true;